// src/app/models/carteira/carteira-produto.ts
export class CarteiraProduto{
    area: number = 0;
    cdBeneficiamento: number =0;
    cdCliente: number =  0;
    cdProduto: number = 0;
    data: string | undefined;
    id: number | undefined;
    nomeBeneficiamento: string = "";
    nomeCliente: string = "";
    nomeProduto: string = "";
    quantidade: number = 0;
    valor: number = 0;
    valorUnitario: number = 0;
}